mula-main/src/main/java/org/mula/finance/activities/habits/list/views/CheckmarkPanelView.kt
/* Mula */

package org.mula.finance.activities.habits.list.views

import android.content.*
import com.google.auto.factory.*
import org.mula.finance.core.models.Checkmark.*
import org.mula.androidbase.activities.ActivityContext
import org.mula.finance.core.models.Timestamp
import org.mula.finance.core.preferences.Preferences
import org.mula.finance.core.utils.DateUtils

@AutoFactory
class CheckmarkPanelView(
        @Provided @ActivityContext context: Context,
        @Provided preferences: Preferences,
        @Provided private val buttonFactory: CheckmarkButtonViewFactory
) : ButtonPanelView<CheckmarkButtonView>(context, preferences) {

    var values = IntArray(0)
        set(values) {
            field = values
            setupButtons()
        }

    var color = 0
        set(value) {
            field = value
            setupButtons()
        }

    var onToggle: (Timestamp) -> Unit = {}
        set(value) {
            field = value
            setupButtons()
        }

    override fun createButton(): CheckmarkButtonView = buttonFactory.create()

    @Synchronized
    override fun setupButtons() {
        val today = DateUtils.getToday()

        buttons.forEachIndexed { index, button ->
            val timestamp = today.minus(index + dataOffset)
            button.value = when {
                index + dataOffset < values.size -> values[index + dataOffset]
                else -> UNCHECKED
            }
            button.color = color
            button.onToggle = { onToggle(timestamp) }
        }
    }
}
